temperature formula fix

+++ lib/trmnl/include/api_types.h
@@ -59,7 +59,7 @@ struct ApiDisplayInputs
   int batteryCurrent;
   int currentBatteryCapacity;
   int maxBatteryCapacity;
-  int batteryTemperature;
+  float batteryTemperature;
   int stateOfCharge;
   int stateOfHealth;
 #endif
